HPOS Feature: Revert to one official feature
In #38993, the `custom_order_tables` feature was split into two
separate features, one for choosing which tables to use, and one for
toggling data synchronization. My understanding is that this was done
in order to have both of those settings appear on the Features settings
page, instead of having to create a separate page for other
HPOS-specific settings. However, this added some extra complexity,
which then led to #39250 causing a regression where extensions that
declared compatibility with HPOS were no longer recognized as
compatible.

With this PR, we're reverting back to only having one "feature" for
HPOS, but introducing a way to add additional setting controls to the
Features settings page.